changed segmentation volume local to segmentation volume
